Update _getTarget function
This minor change allows to insert a tooltip and its target programmatically without the error of parentNode being null.
Originally, without the line 458, when making the conditional in line 459 ( if(parentNode.nodeType === 11) ) from the get go, if the px-tooltip was created using document.createElement(px-tooltip), then the code will run and it will break at line 459 because parentNode is null since its target hasn't been specified. If we first ask for the existence of the object parentNode, if this is null, this skips the condition altogether along with the error.